+++ src/keys.rs
@@ -6,7 +6,9 @@ use alloc::vec::Vec;
 #[derive(Debug, PartialEq)]
 /// Versions associated with a key, used in PASETO.
 pub enum Version {
+    /// Keys for version 2.
     V2,
+    /// Keys for version 4.
     V4,
 }